refactor(core): merge contracts into pack; delete old shape
  * Merge openrange.core.contracts into openrange.core.pack: the Pack /
    Builder / TaskFamily protocols + every wire dataclass now live in
    one module, alongside the PackRegistry / PACKS that already lived
    there. With the legacy Pack ABC gone, PackRegistry isinstance check
    works directly. No more workaround in runtime.py / __main__.py;
    both call PACKS.resolve() and let it propagate.
  * Delete the legacy parallel-path modules whose live consumers all
    migrated in Phases 2 and 3:
      core/admission.py, core/builder.py, core/builder_protocol.py,
      core/graph.py, core/snapshot.py, core/runtime_backing.py,
      core/runtime_helpers.py, core/backings.py, core/manifest.py.
  * Rewrite core/__init__.py and openrange/__init__.py to export ONLY
    the new shape. Drop dead names (Entrypoint, Task, BUILDERS,
    BuildContext, WorldSchema, NodeType, EdgeType, GraphConstraint,
    LineageNode, PackRef, PackSource, Ref, RuntimeArtifact, RuntimeBundle,
    CheckScript, ValidationError, AdmissionReport, AdmissionResult,
    verifier_from_source, admission_state_from_source, build, evolve)
    and surface the new public API (Backing, BuildEvent, BuildResult,
    EpisodeReportLike, EpisodeResult, FeasibilityVerdict, LLMBackendLike,
    PackPrior, RuntimeHandle, TaskFamily, TaskSeed, TaskSpec, admit,
    snapshot_to_dict, validate_task_bindings, distill).
  * Re-export Edge / Node / WorldGraph / Visibility / Role / Ontology /
    AttrSpec / NodeKind / EdgeKind / GraphPatch / Issue / validate /
    apply_patch from openrange.world_ir at the top-level openrange
    package -- these still belong to the public API.
  * Delete tests/test_v1_episode.py: every test in the module is
    skipped at the module level (Round 3 done; behaviors live in the
    new test suite).
