+++ packs/cyber_webapp/cyber_webapp/families/pentest.py
@@ -23,7 +23,7 @@ from __future__ import annotations
 from collections.abc import Mapping, Sequence
 from typing import TYPE_CHECKING, Any, cast
 
-from openrange.core.contracts import (
+from openrange.core.pack import (
     EpisodeReportLike,
     EpisodeResult,
     FeasibilityVerdict,
