+++ src/openrange/__init__.py
@@ -7,54 +7,41 @@ from openrange.agent_backend import (
     StrandsAgentBackend,
 )
 from openrange.core import (
-    BUILDERS,
     PACKS,
     ActorTurn,
     AdmissionError,
     AdmissionFailure,
-    AdmissionReport,
-    AdmissionResult,
-    BuildContext,
+    Backing,
     Builder,
-    BuilderError,
-    BuilderRegistry,
-    BuildFailed,
-    BuildState,
-    CheckScript,
+    BuildEvent,
+    BuildResult,
     CurriculumPolicy,
     Direction,
-    Edge,
-    EdgeType,
-    Entrypoint,
-    GraphConstraint,
-    LineageNode,
+    EpisodeReportLike,
+    EpisodeResult,
+    FeasibilityVerdict,
+    LLMBackendLike,
     Manifest,
     ManifestError,
     Mutation,
-    Node,
-    NodeType,
     OpenRangeError,
     Pack,
     PackError,
-    PackRef,
+    PackPrior,
     PackRegistry,
-    PackSource,
-    Ref,
-    RuntimeArtifact,
-    RuntimeBundle,
+    RuntimeHandle,
     Snapshot,
     SnapshotStore,
     StoreError,
-    Task,
-    ValidationError,
-    WorldGraph,
-    WorldSchema,
-    admission_state_from_source,
+    TaskFamily,
+    TaskSeed,
+    TaskSpec,
+    admit,
     auto_evolve,
-    build,
     direction_from_reports,
-    evolve,
-    verifier_from_source,
+    distill,
+    snapshot_to_dict,
+    validate_task_bindings,
 )
 from openrange.core.episode import (
     AdvanceRequest,
@@ -92,86 +79,103 @@ from openrange.runtime import (
     OpenRangeRun,
     RunConfig,
 )
+from openrange.world_ir import (
+    AttrSpec,
+    AttrType,
+    Edge,
+    EdgeKind,
+    GraphPatch,
+    Issue,
+    Node,
+    NodeKind,
+    Ontology,
+    Role,
+    Visibility,
+    WorldGraph,
+    apply_patch,
+    validate,
+)
 
 __all__ = [
+    "ActorTurn",
     "AdmissionError",
     "AdmissionFailure",
-    "AdmissionReport",
-    "AdmissionResult",
-    "ActorTurn",
     "AdvanceRequest",
     "AgentBackend",
     "AgentBackendError",
+    "AgentNPC",
     "AgentTurn",
-    "BUILDERS",
-    "BuildContext",
-    "BuildFailed",
-    "BuildState",
+    "AttrSpec",
+    "AttrType",
+    "Backing",
+    "BuildEvent",
+    "BuildResult",
     "Builder",
-    "BuilderError",
-    "BuilderRegistry",
     "CODEX_DEFAULT_MODEL",
-    "CheckScript",
     "CodexAgentBackend",
     "CodexBackend",
     "CurriculumPolicy",
     "DashboardServerHandle",
     "Direction",
     "Edge",
-    "EdgeType",
-    "Entrypoint",
+    "EdgeKind",
     "EpisodeCheckpoint",
     "EpisodeError",
     "EpisodeHandle",
     "EpisodeReport",
+    "EpisodeReportLike",
+    "EpisodeResult",
     "EpisodeRuntimeError",
     "EpisodeService",
     "EpisodeUpdate",
-    "GraphConstraint",
+    "FeasibilityVerdict",
+    "GraphPatch",
+    "Issue",
     "LLMBackend",
     "LLMBackendError",
+    "LLMBackendLike",
     "LLMError",
     "LLMRequest",
     "LLMRequestError",
     "LLMResult",
-    "LineageNode",
     "Manifest",
     "ManifestError",
     "Mutation",
     "NPC",
+    "NPCS",
     "NPCError",
     "NPCRegistry",
-    "NPCS",
-    "AgentNPC",
     "Node",
-    "NodeType",
+    "NodeKind",
     "Observation",
+    "Ontology",
     "OpenRangeError",
     "OpenRangeRun",
     "PACKS",
     "Pack",
     "PackError",
-    "PackRef",
+    "PackPrior",
     "PackRegistry",
-    "PackSource",
-    "Ref",
+    "Role",
     "RunConfig",
-    "RuntimeArtifact",
-    "RuntimeBundle",
+    "RuntimeHandle",
     "Snapshot",
     "SnapshotStore",
     "StoreError",
     "StrandsAgentBackend",
-    "Task",
+    "TaskFamily",
+    "TaskSeed",
+    "TaskSpec",
     "TickRequest",
     "TickResult",
-    "ValidationError",
+    "Visibility",
     "WorldGraph",
-    "WorldSchema",
-    "admission_state_from_source",
+    "admit",
+    "apply_patch",
     "auto_evolve",
-    "build",
     "direction_from_reports",
-    "evolve",
-    "verifier_from_source",
+    "distill",
+    "snapshot_to_dict",
+    "validate",
+    "validate_task_bindings",
 ]
