+++ src/openrange/core/__init__.py
@@ -1,73 +1,46 @@
 """OpenRange core public API."""
 
-# Side-effect: registers built-in backings (HTTP) into RUNTIME_BACKINGS.
-import openrange.core.backings as _backings  # noqa: F401, E402
-from openrange.core.admission import (
+from openrange.core.admit import (
     AdmissionFailure,
-    AdmissionReport,
-    AdmissionResult,
-    BuildFailed,
+    BuildEvent,
+    Snapshot,
     admit,
-)
-from openrange.core.builder import (
-    BuildContext,
-    BuildState,
-    build,
-    evolve,
-)
-from openrange.core.builder_protocol import (
-    BUILDERS,
-    Builder,
-    BuilderError,
-    BuilderRegistry,
+    snapshot_to_dict,
+    validate_task_bindings,
 )
 from openrange.core.curriculum import (
     CurriculumPolicy,
     Direction,
-    Mutation,
     auto_evolve,
     direction_from_reports,
 )
+from openrange.core.distill import distill
 from openrange.core.errors import (
     AdmissionError,
+    EpisodeRuntimeError,
     ManifestError,
     OpenRangeError,
     PackError,
     StoreError,
 )
-from openrange.core.graph import (
-    CheckScript,
-    Edge,
-    EdgeType,
-    GraphConstraint,
-    Node,
-    NodeType,
-    Ref,
-    RuntimeArtifact,
-    RuntimeBundle,
-    ValidationError,
-    WorldGraph,
-    WorldSchema,
-)
-from openrange.core.manifest import (
-    Manifest,
-    PackRef,
-    PackSource,
-)
 from openrange.core.pack import (
     PACKS,
-    Entrypoint,
+    Backing,
+    Builder,
+    BuildResult,
+    EpisodeReportLike,
+    EpisodeResult,
+    FeasibilityVerdict,
+    LLMBackendLike,
+    Manifest,
+    Mutation,
     Pack,
+    PackPrior,
     PackRegistry,
-    Task,
-    admission_state_from_source,
-    verifier_from_source,
-)
-from openrange.core.snapshot import (
-    LineageNode,
-    Snapshot,
-    stable_json,
-    task_from_mapping,
+    RuntimeHandle,
+    TaskFamily,
+    TaskSeed,
+    TaskSpec,
 )
 from openrange.core.store import SnapshotStore
 from openrange.core.turn import ActorTurn
@@ -77,51 +50,36 @@ __all__ = [
     "ActorTurn",
     "AdmissionError",
     "AdmissionFailure",
-    "AdmissionReport",
-    "AdmissionResult",
-    "BUILDERS",
-    "BuildContext",
-    "BuildFailed",
-    "BuildState",
+    "Backing",
+    "BuildEvent",
+    "BuildResult",
     "Builder",
-    "BuilderError",
-    "BuilderRegistry",
-    "CheckScript",
     "CurriculumPolicy",
     "Direction",
-    "Edge",
-    "EdgeType",
-    "Entrypoint",
-    "GraphConstraint",
-    "LineageNode",
+    "EpisodeReportLike",
+    "EpisodeResult",
+    "EpisodeRuntimeError",
+    "FeasibilityVerdict",
+    "LLMBackendLike",
     "Manifest",
     "ManifestError",
     "Mutation",
-    "Node",
-    "NodeType",
     "OpenRangeError",
     "Pack",
     "PackError",
-    "PackRef",
+    "PackPrior",
     "PackRegistry",
-    "PackSource",
-    "Ref",
-    "RuntimeArtifact",
-    "RuntimeBundle",
+    "RuntimeHandle",
     "Snapshot",
     "SnapshotStore",
     "StoreError",
-    "Task",
-    "ValidationError",
-    "WorldGraph",
-    "WorldSchema",
-    "admission_state_from_source",
+    "TaskFamily",
+    "TaskSeed",
+    "TaskSpec",
     "admit",
     "auto_evolve",
-    "build",
     "direction_from_reports",
-    "evolve",
-    "stable_json",
-    "task_from_mapping",
-    "verifier_from_source",
+    "distill",
+    "snapshot_to_dict",
+    "validate_task_bindings",
 ]
